requirements.txt
# Discord Bot Dependencies - Railway Deployment
# Install with: pip install -r requirements.txt

# Core Discord library
discord.py==2.6.0

# HTTP client (automatically installed with discord.py, but specified for clarity)
aiohttp>=3.7.4,<4.0.0

# Additional useful packages for bot development
python-dotenv==1.0.0  # For environment variable management
asyncio-throttle==1.0.2  # For rate limiting
colorama==0.4.6  # For colored terminal output

# Railway deployment packages
gunicorn==21.2.0  # WSGI HTTP Server for Railway
psycopg2-binary==2.9.9  # PostgreSQL adapter (if using Railway PostgreSQL)
redis==5.0.1  # Redis client (if using Railway Redis)

# Logging and monitoring
structlog==23.2.0  # Structured logging
sentry-sdk==1.38.0  # Error tracking (optional)

# Utilities
requests==2.31.0  # HTTP requests
python-dateutil==2.8.2  # Date utilities
pytz==2023.3  # Timezone handling
